+++ LoginRadiusSDK/LoginRadiusSDK/Entity/APIs Entity/LoginRadiusAccountEntity.cs
@@ -9,12 +9,27 @@ using LoginRadiusSDK.Models.UserProfile;
 using LoginRadiusSDK.Utility;
 namespace LoginRadiusSDK.Entity
 {
+<<<<<<< HEAD
     public class LoginRadiusAccountEntity : LoginRadiusEntityBase
+=======
+    public class LoginRadiusAccountEntity : LoginRadiusServerEntity
+>>>>>>> master
     {
         readonly LoginRadiusArgumentValidator _validate=new LoginRadiusArgumentValidator();
         private readonly LoginRadiusObject _object = new LoginRadiusObject("account");
         private ArrayList _valuesToCheck;
 
+<<<<<<< HEAD
+=======
+        public LoginRadiusAccountEntity()
+        {
+        }
+        public LoginRadiusAccountEntity(string apikey, string apisecret)
+            : base(apikey,apisecret)
+        {
+        }
+
+>>>>>>> master
         /// <summary>
         /// Method which is used to link a social account with a specified provider user account.
         /// </summary>