+++ LoginRadiusSDK/LoginRadiusSDK/Entity/APIs Entity/LoginRadiusCustomObjectEntity.cs
@@ -9,12 +9,27 @@ using LoginRadiusSDK.Utility.Serialization;
 
 namespace LoginRadiusSDK.Entity
 {
+<<<<<<< HEAD
     public class LoginRadiusCustomObjectEntity : LoginRadiusEntityBase
+=======
+    public class LoginRadiusCustomObjectEntity : LoginRadiusServerEntity
+>>>>>>> master
     {
         private readonly LoginRadiusObject _object = new LoginRadiusObject("user/customObject");
         private readonly LoginRadiusArgumentValidator _validate = new LoginRadiusArgumentValidator();
         private ArrayList _valuesToCheck;
 
+<<<<<<< HEAD
+=======
+        public LoginRadiusCustomObjectEntity()
+        {
+        }
+        public LoginRadiusCustomObjectEntity(string apikey, string apisecret)
+            : base(apikey,apisecret)
+        {
+        }
+
+>>>>>>> master
         /// <summary>
         /// Method is used to write information in JSON format to the custom object for the specified account.
         /// </summary>