+++ LoginRadiusSDK/LoginRadiusSDK/Entity/LoginRadiusCallback.cs
@@ -1,6 +1,10 @@
 ﻿using System;
 using System.Web;
 using LoginRadiusSDK.Exception;
+<<<<<<< HEAD
+=======
+using LoginRadiusSDK.Models;
+>>>>>>> master
 using LoginRadiusSDK.Utility.Http;
 using LoginRadiusSDK.Utility.Serialization;
 
@@ -10,10 +14,27 @@ namespace LoginRadiusSDK.Entity
     /// <summary>
     /// The LoginRadius callback is used to request loginradius api when user  has successfully logged in the preferred provider.
     /// </summary>
+<<<<<<< HEAD
     public class LoginRadiusCallback
     {
         readonly HttpClient _client = new HttpClient();
         string accesstokenendpoint = "api/v2/access_token?token={0}&secret={1}";
+=======
+    public class LoginRadiusCallback : LoginRadiusV2EntityBase
+    {  
+        private readonly LoginRadiusObject _object = new LoginRadiusObject("access_token");
+        private string _token;
+        public LoginRadiusCallback(string token = null)
+        {
+            _token = token;
+        }
+
+        public LoginRadiusCallback(string apikey, string apisecret, string token = null)
+            : base(apikey,apisecret)
+        {
+            _token = token;
+        }
+>>>>>>> master
 
         /// <summary>
         /// To get LoginRadius callback for loginradius api.
@@ -22,7 +43,16 @@ namespace LoginRadiusSDK.Entity
         {
             get
             {
+<<<<<<< HEAD
                 return !string.IsNullOrEmpty(HttpContext.Current.Request.Form["token"]);
+=======
+                if (_token != null)
+                {
+                    return true;
+                }
+                _token = HttpContext.Current.Request.Form["token"];
+                return _token != null;
+>>>>>>> master
             }
         }
 
@@ -35,7 +65,11 @@ namespace LoginRadiusSDK.Entity
             {
                 if (IsCallback)
                 {
+<<<<<<< HEAD
                     return HttpContext.Current.Request.Form["token"];
+=======
+                    return _token;
+>>>>>>> master
                 }
                 else
                 {
@@ -43,10 +77,15 @@ namespace LoginRadiusSDK.Entity
                 }
             }
         }
+<<<<<<< HEAD
+=======
+        
+>>>>>>> master
 
         /// <summary>
         /// The Access Token API is used to get the LoginRadius access token after authentication. It will be valid for the specific duration of time specified in the response.
         /// </summary>
+<<<<<<< HEAD
         /// <param name="secret">A unique secret key, which is provided by LoginRadius on your Site manage page</param>
         /// <returns></returns>
         public LoginRadiusAccessToken GetAccessToken(string secret)
@@ -70,6 +109,57 @@ namespace LoginRadiusSDK.Entity
             {
                 throw new ArgumentException("Secret is not valid format (Guid) or null", "secret");
             }
+=======
+        /// <returns></returns>
+        public LoginRadiusAccessToken GetAccessToken()
+        {
+            if (IsCallback)
+            {
+                var getParams = new HttpRequestParameter
+                {
+                    {"token", RequestToken}
+                };
+                var response = Get(_object, getParams);
+                return response.Deserialize<LoginRadiusAccessToken>();
+            }
+            throw new LoginRadiusException("Request token is not valid or does not exist.");
+        }
+
+        /// <summary>
+        /// API is used to validate access_token, check it is valid, expired or active.
+        /// </summary>
+        /// <returns></returns>
+        public LoginRadiusAccessToken ValidateAccessToken(string accessToken)
+        {
+            var getParams = new HttpRequestParameter
+            {
+                {"access_token", accessToken}
+            };
+            var response = Get(_object.ChildObject("Validate"), getParams);
+            return response.Deserialize<LoginRadiusAccessToken>();
+        }
+        public LoginRadiusAccessToken ValidateAccessToken(Guid accessToken)
+        {
+            return ValidateAccessToken(accessToken.ToString());
+        }
+
+        /// <summary>
+        ///  API is used to invalidate access token, means expiring token. After this API call passed access_token no longer be active and will not accepted by LoginRadius APIs.
+        /// </summary>
+        /// <returns></returns>
+        public LoginRadiusPostResponse InValidateAccessToken(string accessToken)
+        {
+            var getParams = new HttpRequestParameter
+            {
+                {"access_token", accessToken}
+            };
+            var response = Get(_object.ChildObject("invalidate"), getParams);
+            return response.Deserialize<LoginRadiusPostResponse>();
+        }
+        public LoginRadiusPostResponse InValidateAccessToken(Guid accessToken)
+        {
+            return InValidateAccessToken(accessToken.ToString());
+>>>>>>> master
         }
     }
 }