+++ LoginRadiusSDK/LoginRadiusSDK/Models/Object/UserRegistrationModel.cs
@@ -1,0 +1,8 @@
+﻿namespace LoginRadiusSDK.Models.Object
+{
+    public class UserRegistrationModel:User
+    {
+        public string EmailVerificationUrl { get; set; }
+        public string Template { get; set; }
+    }
+}